refactor: Align peek view detail components
