+++ web/src/components/table/peek/peek-observation-detail.tsx
@@ -4,7 +4,7 @@ import { Trace } from "@/src/components/trace/Trace";
 import { Skeleton } from "@/src/components/ui/skeleton";
 import { useRouter } from "next/router";
 
-export const PeekViewObservationDetail = ({
+const PeekViewObservationDetail = ({
   peekId,
   trace,
 }: {
